requirements.txt
Django==4.0.4
django-admin-sortable2==1.0.4
django-compressor==3.1
django-cors-headers==3.12.0
django-libsass==0.9
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
sentry-sdk==1.5.10
sorl-thumbnail==12.8.0
